fix: implement ExampleModel in src instead of test
